move common user role logic in trait
