+++ app/Http/Controllers/TransactionsController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Models\Status;
 use App\Models\Transaction;
 use App\Models\User;
+use App\Traits\AbilityTrait;
 use Exception;
 use Illuminate\Foundation\Http\FormRequest;
 use Illuminate\Http\Client\Response;
@@ -13,51 +14,10 @@ use Illuminate\Support\Facades\Log;
 
 class TransactionsController extends Controller
 {
-    /**
-     * Checks if request sent from admin
-     *
-     * @param string $fullToken
-     * @return boolean
-     */
-    private function checkIfAdmin($fullToken)
-    {
-        [$bearer_id, $token] = explode('|', $fullToken, 2);
-
-        $tokenId = explode(' ', $bearer_id, 2)[1];
-        /**
-         * @var User
-         */
-        $user = User::select(['*', 'personal_access_tokens.id as p_id', 'users.id as id'])
-            ->join('personal_access_tokens', 'users.id', 'personal_access_tokens.tokenable_id')
-            ->firstWhere('personal_access_tokens.id', $tokenId);
-
-        return $user->hasRole('admin');
-    }
+    use AbilityTrait;
 
     /**
-     * Validate user role
-     *
-     * @param FormRequest $request
-     * @param User $authUser
-     * @return void
-     */
-    protected function validateTokenAndUserRole(FormRequest $request, $authUser = null)
-    {
-        if ($authUser) {
-            $isAdmin = $authUser->hasRole('admin');
-        } else {
-            $fullToken = $request->header('Authorization') ?? null;
-
-            if (!$fullToken) throw new Exception('No token provided.');
-
-            $isAdmin = $this->checkIfAdmin($fullToken);
-
-            if (!$isAdmin) throw new Exception('Admins only can add transactions');
-        }
-    }
-
-    /**
-     * Get Tranasction status based on current time & due on date
+     * Get Transaction status based on current time & due on date
      *
      * @param Carbon $dueOnDate
      * @return Status
@@ -135,15 +95,5 @@ class TransactionsController extends Controller
             Log::error($th->getMessage());
             return response()->json(['message' => $th->getMessage()], 400);
         }
-    }
-
-    public function allTransactions(FormRequest $request)
-    {
-
-    }
-
-    public function userTransactions(FormRequest $request)
-    {
-
     }
 }
